backend/requirements.txt
alembic==1.16.5
alembic-postgresql-enum==1.8.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
boto3==1.40.21
botocore==1.40.21
botocore-stubs==1.40.21
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.6
cryptography==45.0.7
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==6.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
responses==0.25.8
ruff==0.12.11
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.3
types-awscrt==0.27.6
types-boto3==1.40.21
types-boto3-s3==1.40.0
types-s3transfer==0.13.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
xmltodict==0.15.1
